BMP581 and GPS mods
This was the first attempt at adding code for the BMP581.  It is close but not quite there.

The only mod to the GPS was to set the minimum time it is enabled to 30000 millis (30 sec).  This was done so that the system could be put to sleep for long intervals, woken up, and get a valid GPS before going back to sleep.  The data sheet for the SAM-M8Q lists the TTFF for cold start as 29 secs.